+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 Version `2.0.0` is considered as the first "ready for use" release and changes before this point are not documented.
 
+## `2.1.2`
+### Added
+- Extra trace-level logging with FFT frequency and sampling information
+
+### Fixed
+- Crash when FFT trim was above Nyquist frequency
+
 ## `2.1.1`
 ### Added
 - Separate error messages for missing column or missing header in data files